fix(ctx_read,ci): restore the diff baseline, split the file, refresh contracts
Four red checks on this PR, four separate causes.

Test (ubuntu/windows) — `phase2_read_diff_after_edit` failed with
"[no cached version for diff — use mode=full first, then diff on re-read]".
The staleness guard invalidates a cache entry whose mtime/hash no longer match
the file on disk, which is right for every mode except `diff`: there the stale
entry *is* the baseline the delta is computed against, so an edit destroyed
exactly the thing the next `diff` read needed. `diff` is now exempt from that
invalidation; every other mode still refuses content that no longer matches
disk. The test only passed before #1584 by accident — the old MCP path answered
"[WARNING: unknown mode 'diff']" and returned the whole file, which happened to
contain the string the assertion looked for.

Clippy — the job's `loc-gate.sh` step: ctx_read.rs had grown to 1579 lines
(> 1500). Five leaf helpers (`record_attribution_result`, `apply_verdict`,
`auto_degrade_read_mode`, `extract_file_summary`, `task_intent_steers_read`)
move to `ctx_read_helpers.rs` behind the existing `#[path] mod` pattern the
file already uses for `engine`/`image`/`window`. No behaviour change; the
inline tests reach them unchanged via `super::`. ctx_read.rs is now 1496 lines.
The allowlist is untouched — it says shrink, never extend.

Delivery + Security Verification — self-inflicted: the previous commit dropped
a trailing blank line from `.github/workflows/release.yml` to satisfy the
whitespace gate, and `docs/contracts/delivery-evidence-v1.json` pins that
file's digest twice (as `release.workflow` and as the `engine-release` publish
channel). Both refreshed, contract re-emitted as canonical JSON.

cargo-deny — `chacha20 0.10.1` was yanked upstream (reached via `rand 0.10.2`
from both `rmcp` and `lopdf`). Lockfile bumped to 0.10.2; `cargo deny check
advisories` is ok. Unlike the arrayref case documented in deny.toml, this yank
resolves upward with no supply-chain concern, so no ignore entry is added.

Verification: cargo test --lib → 10497 passed, 0 failed, 22 ignored.
cargo test --test power_user_worksession → 24 passed, 0 failed.
cargo clippy --lib --all-features clean, cargo fmt --check clean,
loc-gate.sh ok, verify-delivery-evidence.py exit 0,
cargo deny check advisories ok.
